refactor: move generators to proxy_utils
